+++ src/toolkits/drawing_classifier/dc_data_iterator.hpp
@@ -168,9 +168,9 @@ class simple_data_iterator : public data_iterator {
       const gl_sarray& targets, std::vector<std::string> expected_class_labels);
 
   gl_sframe data_;
-  const int target_index_;
-  const int predictions_index_;
-  const int feature_index_;
+  const size_t target_index_;
+  const size_t predictions_index_;
+  const size_t feature_index_;
   const bool repeat_;
   const bool shuffle_;
